에세이 디테일 api 공통화
