+++ desktop_app/renderer/shared/api/index.ts
@@ -1,2 +1,3 @@
 export { default as AxiosInstance } from "./axiosInstance";
 export { getUserInfo } from "./getUserInfo";
+export { getEssayDetail } from "./essay";
